YT-28704: Save a JNBD snapshot without materializing the block map cut
`TakeSnapshot` materialized the whole block map cut — 16 bytes per used block — and it had to outlive the flush barrier, since a dirty block's stored location is only known once flushed. That vector was the entire gap between save and restore peak memory.

`IBlockMap` exposes the cut as an open scan instead:

```cpp
void BeginSnapshot();                            // arm copy-on-write, fixing the point in time
TBlockMapSnapshot ScanSnapshotPart(begin, end);  // the cut over an index range, as often as asked
void EndSnapshot();                              // disarm, drop the stash
```

A save holds one cut open across the barrier and walks it twice in parts: once to register the dirty versions to wait for, once to emit each part as a write batch. Nothing per-block is retained in between, so the peak is one part.

### Why a re-scan is faithful

A writer stashes a block's pre-snapshot value before storing the CoW-marked new one, both under `WriteLock_`. So a slot found marked already has its stash, and a bare slot has not moved since the flip. `ScanSnapshotPart` collects the marked positions while scanning lock-free, then resolves them under one lock per part.

### Measured

Isolated server, 1 TiB device restored from a live snapshot, 178'297'632 used blocks, three saves per build:

| | save peak | save time (mean of 3) |
|---|---|---|
| trunk | 5.96–6.04 GB | 66.3 s |
| this | **3.19–3.21 GB** | 68.5 s |

### The trade

The CoW window used to be just the scan; it now spans the barrier and the emit loop. The stash holds one entry per distinct block overwritten while the cut is open, so memory follows write volume rather than device size — worst case, a device rewritten during a save, is worse than before.
commit_hash:0f67b6653b174ef4f0e23ec89a8d08528de11384

+++ yt/yt/server/lib/nbd/journal/config.h
@@ -169,9 +169,9 @@ struct TJournalBlockDeviceConfig
     //! Background compaction of mostly-dead chunks; absent disables it.
     TJournalBlockCompactorConfigPtr BlockCompactor;
 
-    //! How many blocks a snapshot resolves and writes at a time. Caps the save's peak memory, which
-    //! would otherwise scale with the device's block count.
-    i64 SnapshotBlocksPerBatch;
+    //! How wide a block index window a snapshot scans, resolves and writes at a time. Caps the save's
+    //! peak memory, which would otherwise scale with the device's block count.
+    int SnapshotBlocksPerBatch;
 
     REGISTER_YSON_STRUCT(TJournalBlockDeviceConfig);
